speech_to_text/requirements.txt
portaudio19-dev
python3-pyaudio
SpeechRecognition